Give BoredApi as a dependency to ViewModel
