+++ app/src/main/java/com/github/leuludyha/ibdb/webapi/WebApiApplication.kt
@@ -2,11 +2,15 @@ package com.github.leuludyha.ibdb.webapi
 
 import android.app.Application
 import com.github.leuludyha.ibdb.database.*
+import dagger.hilt.android.HiltAndroidApp
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.SupervisorJob
 
 // TODO Modify this application to have a more general one
-class WebApiApplication : Application() {
+@HiltAndroidApp
+open class WebApiApplication : Application() {
+    open fun getBaseUrl() = "https://www.boredapi.com/api/"
+
     // No need to cancel this scope as it'll be torn down with the process
     private val applicationScope = CoroutineScope(SupervisorJob())
